refact and comments addressed
