+++ optimizely/helpers/audience.py
@@ -12,6 +12,7 @@
 # limitations under the License.
 
 from . import condition as condition_helper
+from . import condition_tree_evaluator
 
 
 def is_match(audience, attributes):
@@ -24,7 +25,6 @@ def is_match(audience, attributes):
   Return:
     Boolean representing if user satisfies audience conditions or not.
   """
-  condition_tree_evaluator = condition_helper.ConditionTreeEvaluator()
   custom_attr_condition_evaluator = condition_helper.CustomAttributeConditionEvaluator(
     audience.conditionList, attributes)
