+++ optimizely/helpers/validator.py
@@ -13,6 +13,8 @@
 
 import json
 import jsonschema
+import math
+import numbers
 from six import string_types
 
 from optimizely.user_profile import UserProfile
@@ -186,6 +188,60 @@ def is_attribute_valid(attribute_key, attribute_value):
     return False
 
   if isinstance(attribute_value, string_types) or type(attribute_value) in (int, float, bool):
+    return True
+
+  return False
+
+
+def is_finite_number(value):
+  """ Method to validate if the given value is a number and not one of NAN, INF, -INF
+
+  Args:
+    value: Value to be validated
+
+  Returns:
+    Boolean: True if value is a number and not NAN, INF or -INF else False
+  """
+  if not isinstance(value, (int, float)):
+      # numbers.Integral instead of int to accomodate long integer in python 2
+    return False
+
+  if isinstance(value, bool):
+    # bool is a subclass of int
+    return False
+
+  if isinstance(value, float):
+    if math.isnan(value) or math.isinf(value):
+      return False
+
+  return True
+
+
+def are_values_same_type(first_val, second_val):
+  """ Method to verify that both values belong to same type. Float and integer are
+  considered as same type.
+
+  Args:
+    first_val: Value to validate
+    second_Val: Value to validate
+
+  Returns:
+    Boolean: True if both values belong to same type. Otherwise False
+  """
+
+  first_val_type = type(first_val)
+  second_val_type = type(second_val)
+
+  # use isinstance to accomodate Python 2 unicode and str types
+  if isinstance(first_val, string_types) and isinstance(second_val, string_types):
+    return True
+
+  # Compare types if one of the values is bool because bool is a subclass on Integer
+  if isinstance(first_val, bool) or isinstance(second_val, bool):
+    return first_val_type == second_val_type
+
+  # Treat ints and floats as same type
+  if isinstance(first_val, (numbers.Integral, float)) and isinstance(second_val, (numbers.Integral, float)):
     return True
 
   return False
